[Refactor] import recommended assertions rules
